requirements.txt

langchain
langchain-community

google-generativeai

streamlit>=1.32
python-dotenv>=1.0
google-generativeai>=0.7.0
